Added currencies to ClaimType/UnitType.

+++ include/opentxs/core/Types.hpp
@@ -13,6 +13,6 @@ namespace opentxs
 namespace core
 {
 enum class AddressType : std::uint8_t;
-enum class UnitType : std::uint8_t;
+enum class UnitType : std::uint32_t;
 }  // namespace core
 }  // namespace opentxs
